requirements.txt
# Django и основные зависимости
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-environ>=0.10.0

# База данных (PostgreSQL рекомендуется для production)
psycopg2-binary>=2.9.0

# Аутентификация и безопасность
djangorestframework-simplejwt>=5.2.0

# Задачи по расписанию (Celery)
celery>=5.3.0
redis>=4.5.0

# Геопространственные вычисления (для проверки уникальности по радиусу)
geopy>=2.3.0

# HTTP запросы (для геокодирования)
requests>=2.31.0

# Утилиты
python-dateutil>=2.8.0
pytz>=2023.3

# Разработка и тестирование
pytest>=7.4.0
pytest-django>=4.5.0
black>=23.7.0
flake8>=6.1.0

